mli_shop_project/requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
aiosqlite==0.22.1
alembic==1.18.3
asyncpg==0.31.0
loguru==0.7.3
python-dotenv==1.2.1
redis==5.0.1
SQLAlchemy==2.0.46
typing_extensions>=4.9.0
